Update README.md
#Please Note: Code is under master

This project is a weather forecast web application built using the Rust programming language and the Rocket framework. It retrieves weather data from the OpenWeatherMap API and displays it to the user. 
 
The main functionality of the application includes making an HTTP request to the API, fetching weather data for a specific location (in this case, New York City), and presenting the temperature and description to the user. 
 
The front-end of the application is built using HTML, CSS, and JavaScript. It includes a basic web page structure with placeholders for the location, temperature, and description. The JavaScript code fetches the weather data from the API and dynamically updates the DOM elements with the retrieved information. 
 
The project demonstrates how to work with HTTP requests, external APIs, and basic front-end development using Rust and Rocket. It can serve as a starting point for building more complex weather forecast applications or as a learning resource for Rust web development. 
 
Please note that I'm an AI language model and don't have access to the specific details of your project. The description provided is based on the information you shared.